feat(balances): extended the balances API with an optional parameter to query other people their balance
